Make the proper data types in CacheManager
Use uint16_t instead of uint8_t in order to avoid
overflow. Since the time to time overflow happened and causes
PTU absence during checking ignition triggers.

Closes-Bug: APPLINK-18146

+++ src/components/policy/include/policy/pt_representation.h
@@ -105,7 +105,7 @@ class PTRepresentation {
      * @param current value in days after epoch
      * @return value in days
      */
-    virtual int DaysBeforeExchange(int current) = 0;
+    virtual int DaysBeforeExchange(uint16_t current) = 0;
 
     /**
      * @brief Increment number of ignition cycles since last exchange by 1
